Improve coverage and avoid duplication
