+++ src/main/java/com/powsybl/openloadflow/dc/equations/ClosedBranchSide1DcFlowEquationTerm.java
@@ -6,7 +6,6 @@
  */
 package com.powsybl.openloadflow.dc.equations;
 
-import com.powsybl.math.matrix.DenseMatrix;
 import com.powsybl.openloadflow.equations.Variable;
 import com.powsybl.openloadflow.equations.VariableSet;
 import com.powsybl.openloadflow.network.LfBranch;
@@ -38,17 +37,10 @@ public final class ClosedBranchSide1DcFlowEquationTerm extends AbstractClosedBra
         return new ClosedBranchSide1DcFlowEquationTerm(branch, bus1, bus2, variableSet, deriveA1, useTransformerRatio);
     }
 
-    private double calculate(double ph1, double ph2, double a1) {
+    @Override
+    protected double calculate(double ph1, double ph2, double a1) {
         double deltaPhase =  ph2 - ph1 + A2 - a1;
         return -power * deltaPhase;
-    }
-
-    public double calculate(DenseMatrix x, int column) {
-        Objects.requireNonNull(x);
-        double ph1 = x.get(ph1Var.getRow(), column);
-        double ph2 = x.get(ph2Var.getRow(), column);
-        double a1 = a1Var != null && a1Var.isActive() ? x.get(a1Var.getRow(), column) : branch.getPiModel().getA1();
-        return calculate(ph1, ph2, a1);
     }
 
     @Override
@@ -56,7 +48,7 @@ public final class ClosedBranchSide1DcFlowEquationTerm extends AbstractClosedBra
         Objects.requireNonNull(x);
         double ph1 = x[ph1Var.getRow()];
         double ph2 = x[ph2Var.getRow()];
-        double a1 = a1Var != null && a1Var.isActive() ? x[a1Var.getRow()] : branch.getPiModel().getA1();
+        double a1 = getA1(x);
         p1 = calculate(ph1, ph2, a1);
         if (a1Var != null && a1Var.isActive()) {
             rhs = -power * A2;
